fix(server): Remove the old Express server instance
IT's been months since we are running the Fastify server instead of Express.
In this commit, I delete the old instance.
