+++ e2e/dm.en.spec.ts
@@ -112,7 +112,7 @@ describe("[default language - english]", () => {
     const TelegramBotModel = init.TelegramBotModel;
     const TelegramBaseApi = init.TelegramBaseApi;
     const mockTgGetWebHook = initTest.mockTgGetWebHook;
-    const BotServer = init.BotServerNew;
+    const BotServer = init.BotServer;
     const appVersion = init.appVersion;
     const launchTime = init.launchTime;
