+++ e2e/dm.ru.spec.ts
@@ -119,7 +119,7 @@ describe("[russian language]", () => {
     const mockTgGetWebHook = initTest.mockTgGetWebHook;
     const mockTgSetWebHook = initTest.mockTgSetWebHook;
     const mockTgSetCommands = initTest.mockTgSetCommands;
-    const BotServer = init.BotServerNew;
+    const BotServer = init.BotServer;
     const appVersion = init.appVersion;
     const launchTime = init.launchTime;
