+++ e2e/errors.spec.ts
@@ -48,7 +48,7 @@ describe("error cases", () => {
       const mockTgGetWebHook = initTest.mockTgGetWebHook;
       const mockTgSetWebHook = initTest.mockTgSetWebHook;
       const mockTgSetCommands = initTest.mockTgSetCommands;
-      const BotServer = init.BotServerNew;
+      const BotServer = init.BotServer;
       const appVersion = init.appVersion;
       const launchTime = init.launchTime;
