+++ e2e/group.en.spec.ts
@@ -125,7 +125,7 @@ describe("[default language - english]", () => {
     const mockTgSetWebHook = initTest.mockTgSetWebHook;
     const mockTgSetCommands = initTest.mockTgSetCommands;
     const getVoiceConverterInstances = init.getVoiceConverterInstances;
-    const BotServer = init.BotServerNew;
+    const BotServer = init.BotServer;
     const appVersion = init.appVersion;
     const TelegramBaseApi = init.TelegramBaseApi;
     const localhostUrl = init.localhostUrl;
