+++ e2e/group.ru.spec.ts
@@ -130,7 +130,7 @@ describe("[russian language]", () => {
     const mockTgSetWebHook = initTest.mockTgSetWebHook;
     const mockTgSetCommands = initTest.mockTgSetCommands;
     const getVoiceConverterInstances = init.getVoiceConverterInstances;
-    const BotServer = init.BotServerNew;
+    const BotServer = init.BotServer;
     const appVersion = init.appVersion;
     const TelegramBaseApi = init.TelegramBaseApi;
     const localhostUrl = init.localhostUrl;
