+++ e2e/ignore.spec-ignored.ts
@@ -76,7 +76,7 @@ describe("ignore chats", () => {
     const TelegramBotModel = init.TelegramBotModel;
     const TelegramBaseApi = init.TelegramBaseApi;
     const mockTgGetWebHook = initTest.mockTgGetWebHook;
-    const BotServer = init.BotServerNew;
+    const BotServer = init.BotServer;
     const appVersion = init.appVersion;
     const launchTime = init.launchTime;
