+++ package.json
@@ -41,17 +41,16 @@
     "@google-cloud/speech": "4.5.5",
     "@mtcute/node": "0.22.3",
     "@ngrok/ngrok": "1.4.1",
-    "@sentry/node": "9.8.0",
-    "@sentry/profiling-node": "9.8.0",
+    "@sentry/node": "9.11.0",
+    "@sentry/profiling-node": "9.11.0",
     "aws-sdk": "2.1310.0",
     "axios": "1.8.4",
-    "elevenlabs": "1.54.0",
-    "express": "4.21.2",
-    "fastify": "5.2.1",
+    "elevenlabs": "1.56.0",
+    "fastify": "5.2.2",
     "ffmpeg-static": "5.2.0",
     "fluent-ffmpeg": "2.1.3",
     "nanoid": "5.1.5",
-    "newrelic": "12.16.0",
+    "newrelic": "12.17.0",
     "pg": "8.14.1",
     "picocolors": "1.1.1",
     "signal-exit": "4.1.0",
@@ -64,28 +63,27 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "3.3.1",
-    "@eslint/js": "9.23.0",
-    "@types/express": "5.0.1",
+    "@eslint/js": "9.24.0",
     "@types/fluent-ffmpeg": "2.1.27",
-    "@types/node": "22.13.11",
+    "@types/node": "22.14.0",
     "@types/pg": "8.11.11",
-    "@types/supertest": "6.0.2",
-    "@typescript-eslint/eslint-plugin": "8.27.0",
-    "@typescript-eslint/parser": "8.27.0",
-    "@vitest/coverage-istanbul": "3.0.9",
-    "@vitest/coverage-v8": "3.0.9",
-    "@vitest/ui": "3.0.9",
+    "@types/supertest": "6.0.3",
+    "@typescript-eslint/eslint-plugin": "8.29.0",
+    "@typescript-eslint/parser": "8.29.0",
+    "@vitest/coverage-istanbul": "3.1.1",
+    "@vitest/coverage-v8": "3.1.1",
+    "@vitest/ui": "3.1.1",
     "artillery": "2.0.22",
     "artillery-plugin-expect": "2.16.0",
-    "eslint": "9.23.0",
+    "eslint": "9.24.0",
     "eslint-config-prettier": "10.1.1",
     "husky": "9.1.7",
     "lint-staged": "15.5.0",
-    "nock": "14.0.1",
+    "nock": "14.0.2",
     "prettier": "3.5.3",
     "query-string": "9.1.1",
     "supertest": "7.1.0",
-    "vitest": "3.0.9",
+    "vitest": "3.1.1",
     "vitest-sonar-reporter": "2.0.0"
   },
   "lint-staged": {
@@ -110,7 +108,7 @@
     "bot",
     "speech-to-text",
     "recognition",
-    "express",
+    "fastify",
     "speech",
     "voice",
     "text",
