+++ src/server/bot-server.ts
@@ -1,28 +1,28 @@
-import { createServer as createHttp } from "node:http";
-import { createServer as createHttps } from "node:https";
-import express, { type Response, type Express } from "express";
+import { fastify, type FastifyInstance } from "fastify";
 import { Logger } from "../logger/index.js";
-import { sSuffix } from "../text/utils.js";
-import { BotServerBase } from "./bot-server-base.js";
 import { AnalyticsData } from "../analytics/ga/types.js";
-import { flattenPromise } from "../common/helpers.js";
+import { BotServerBase } from "./bot-server-base.js";
+import { initSentryNew, trackAPIHandlersNew } from "../monitoring/sentry.js";
 import { collectAnalytics } from "../analytics/index.js";
+import { sSuffix } from "../text/utils.js";
+import { isFileExist, readFile } from "../files/index.js";
+import { flattenPromise } from "../common/helpers.js";
 import { TgUpdateSchema } from "../telegram/api/types.js";
-import { initSentry } from "../monitoring/sentry.js";
+import { getMB } from "../memory/index.js";
 import {
   type BotServerModel,
   type HealthDto,
-  type ServerStatCore,
+  type NotFoundDto,
   HealthModel,
 } from "./types.js";
-import type { TelegramBotModel } from "../telegram/bot.js";
-import type { HttpsOptions } from "../../certs/index.js";
 import type { VoidPromise } from "../common/types.js";
+import type { HttpsOptions } from "../../certs/index.js";
+import type { TelegramBotModel } from "../telegram/bot.js";
 
 const logger = new Logger("server");
 
 export class BotServer
-  extends BotServerBase<Express>
+  extends BotServerBase<FastifyInstance>
   implements BotServerModel
 {
   constructor(
@@ -31,135 +31,54 @@ export class BotServer
     webhookDoNotWait: boolean,
     httpsOptions?: HttpsOptions,
   ) {
-    super("ExpressJS", port, version, webhookDoNotWait, httpsOptions);
+    super("Fastify", port, version, webhookDoNotWait, httpsOptions);
 
-    initSentry();
+    initSentryNew();
+    trackAPIHandlersNew(this.app);
 
-    this.app.use("/static", express.static("assets/v2"));
-
-    const statusHandler = (
-      res: Response<HealthDto>,
-      db: ServerStatCore | null,
-    ): void => {
-      const status = new HealthModel(
-        this.version,
-        this.isHttps,
-        this.threadId,
-        this.serverName,
-        this.nodeVersion,
-      );
-      if (this.isIdle) {
-        status.setMessage("App is not connected to the Telegram server");
-        res.status(400).send(status.getDto());
-        return;
-      }
-
-      if (!db?.isReady()) {
-        logger.warn("Database is not ready");
-        // TODO fix behavior
-        // const errMessage = "Database is not ready";
-        // logger.error(errMessage, new Error(errMessage));
-        // status.setMessage("Database is not ready");
-        // res.status(400).send(status.getDto());
-        // return;
-      }
-
-      Promise.all(this.bots.map((bot) => bot.getHostLocation()))
-        .then((urls) => {
-          status.setOnline(urls);
-          res.status(200).send(status.getDto());
-        })
-        .catch((err) => {
-          const errMessage = "Unable to get bot urls";
-          logger.error(errMessage, err);
-          status.setMessage(errMessage, true);
-          res.status(400).send(status.getDto());
-        });
-    };
-
-    this.app.get("/health", (_req, res: Response<HealthDto>) => {
-      statusHandler(res, this.stat);
+    this.app.get<{ Reply: string }>("/favicon.ico", async (_req, reply) => {
+      return reply.status(204).type("image/vnd.microsoft.icon").send("");
     });
-    this.app.get("/favicon.ico", (_req, res: Response<string>) => {
-      res.status(204).send("");
-    });
-    this.app.get("/", (_req, res: Response<string>) => {
+
+    this.app.get<{ Reply: string }>("/", async (_req, reply) => {
       logger.info("Received app root request");
-      res.status(200).send("The app is running");
+      return reply.status(200).type("text/plain").send("The app is running");
     });
-    this.app.post("/lifecycle", (req, res: Response<HealthDto>) => {
-      logger.warn("Received app:restart hook from the buddy node", req.body);
-      statusHandler(res, this.stat);
-    });
-  }
 
-  public setBots(bots: TelegramBotModel[] = []): this {
-    super.setBots(bots);
-
-    bots.forEach((bot) => {
-      logger.warn(`Setting up a handler for ${Logger.y(bot.getPath())}`);
-      this.app.post(bot.getPath(":id"), (req, res) => {
-        if (req.params.id !== bot.getId()) {
-          logger.warn(
-            "Wrong route id! Perhaps because of the cache on Telegram side.",
-            {
-              routeId: req.params.id,
-              botId: bot.getId(),
-            },
-          );
-        }
-
-        const analytics = new AnalyticsData(
-          this.version,
-          this.selfUrl,
-          this.threadId,
+    this.app.get<{ Params: { path: string }; Reply: Buffer }>(
+      "/static/:path",
+      async (req, reply) => {
+        const assetPath = new URL(
+          `../../assets/v2/${req.params.path}`,
+          import.meta.url,
         );
-
-        try {
-          const payload = TgUpdateSchema.parse(req.body);
-          logger.debug("Incoming message validated");
-
-          if (this.webhookDoNotWait) {
-            res.sendStatus(200);
-            logger.debug("Webhook response sent");
-          }
-
-          return bot
-            .handleApiMessage(payload, analytics)
-            .catch((err) => {
-              logger.error("Incoming message failed to handle", err);
-            })
-            .finally(() => {
-              if (this.webhookDoNotWait) {
-                return;
-              }
-              res.sendStatus(200);
-              logger.debug("Webhook response sent");
-            });
-        } catch (err) {
-          logger.error("Incoming message failed validation", err);
-          res.sendStatus(200);
+        const isExists = await isFileExist(assetPath);
+        if (!isExists) {
+          return reply.callNotFound();
         }
-      });
+        const buffer = await readFile(assetPath);
+        return reply.type("image/png").send(buffer);
+      },
+    );
 
-      this.app.get(bot.getPath(":id"), (req, _res, next) => {
-        if (req.params.id !== bot.getId()) {
-          logger.warn(
-            "Wrong route id! Perhaps because of the cache on Telegram side.",
-            {
-              routeId: req.params.id,
-              botId: bot.getId(),
-            },
-          );
-        }
-
-        logger.info("Route is enabled");
-        next();
-      });
+    this.app.get<{ Reply: HealthDto }>("/health", async (_req, reply) => {
+      const [status, dto] = await this.getStatusHandler();
+      return reply.status(status).send(dto);
     });
 
-    this.app.use((req, res) => {
-      const err = new Error("Unknown route");
+    this.app.post<{ Reply: HealthDto; Body: Record<string, unknown> }>(
+      "/lifecycle",
+      async (req, reply) => {
+        logger.warn("Received app:restart hook from the buddy node", req.body);
+        const [status, dto] = await this.getStatusHandler();
+        return reply.status(status).send(dto);
+      },
+    );
+
+    this.app.setNotFoundHandler<{ Reply: NotFoundDto }>(async (req, reply) => {
+      const err = new Error("Unknown route", {
+        cause: { path: req.originalUrl },
+      });
       logger.error(`Unknown route ${Logger.y(req.originalUrl)}`, err);
       const analytics = new AnalyticsData(
         this.version,
@@ -167,55 +86,30 @@ export class BotServer
         this.threadId,
       );
 
-      analytics.addError("Unknown route for the host");
+      analytics
+        .addError("Unknown route for the host")
+        .setCommand("/app", "Server route not found");
 
-      return collectAnalytics(
-        analytics.setCommand("/app", "Server route not found"),
-      ).then(() => {
-        res.status(404).send({
-          status: 404,
-          message: "Route not found",
-          error: "Not found",
-        });
+      await collectAnalytics(analytics);
+      return reply.status(404).send({
+        status: 404,
+        message: "Route not found",
+        error: "Not found",
       });
     });
 
-    return this;
-  }
+    this.app.setErrorHandler(async (error, _req, reply) => {
+      logger.error("Router error", error);
+      const analytics = new AnalyticsData(
+        this.version,
+        this.selfUrl,
+        this.threadId,
+      );
 
-  public start(): Promise<VoidPromise> {
-    const isHttps = Boolean(this.httpsOptions);
-    logger.info(`Starting ${Logger.y(sSuffix("http", isHttps))} server`);
+      analytics.addError("Router error").setCommand("/app", "Server error");
 
-    const server = this.httpsOptions
-      ? createHttps(this.httpsOptions, this.app)
-      : createHttp(this.app);
-
-    return new Promise((resolve) => {
-      server.listen(this.port, () => {
-        logger.info(`The bot server is listening on ${Logger.y(this.port)}`);
-        resolve(
-          () =>
-            new Promise((resolveFn, rejectFn) => {
-              logger.warn("Shutting down the server instance");
-              if (this.uptimeDaemon.isRunning) {
-                logger.warn("Stopping the daemon");
-                this.uptimeDaemon.stop();
-              }
-
-              server.close((err) => {
-                if (err) {
-                  logger.error("Unable to stop the bot server", err);
-                  rejectFn(err);
-                  return;
-                }
-
-                logger.warn("The bot server has stopped");
-                resolveFn();
-              });
-            }),
-        );
-      });
+      await collectAnalytics(analytics);
+      return reply.status(500).send({ error: "Something went wrong" });
     });
   }
 
@@ -226,6 +120,128 @@ export class BotServer
     ).then(() => {
       this.isIdle = false;
       logger.info("Instance is successfully set as a hook receiver");
+    });
+  }
+
+  public setBots(bots: TelegramBotModel[] = []): this {
+    super.setBots(bots);
+
+    bots.forEach((bot) => {
+      logger.warn(`Setting up a handler for ${Logger.y(bot.getPath())}`);
+      this.app.post<{ Params: { id: string }; Body: unknown; Reply: "" }>(
+        bot.getPath(":id"),
+        async (req, reply) => {
+          const routeId = req.params.id;
+          const botId = bot.getId();
+          if (routeId !== botId) {
+            logger.warn(
+              "Wrong route id! Perhaps because of the cache on Telegram side.",
+              {
+                routeId,
+                botId,
+                method: req.method,
+                url: req.originalUrl,
+              },
+            );
+          }
+
+          const analytics = new AnalyticsData(
+            this.version,
+            this.selfUrl,
+            this.threadId,
+          );
+
+          try {
+            const payload = TgUpdateSchema.parse(req.body);
+            logger.debug("Incoming message validated");
+
+            if (this.webhookDoNotWait) {
+              logger.debug("Webhook response sent");
+              return reply.status(200).send("");
+            }
+
+            return bot
+              .handleApiMessage(payload, analytics)
+              .catch((err) => {
+                logger.error("Incoming message failed to handle", err);
+              })
+              .then(() => {
+                if (this.webhookDoNotWait) {
+                  return;
+                }
+                logger.debug("Webhook response sent");
+                return reply.status(200).send("");
+              });
+          } catch (err) {
+            logger.error("Incoming message failed validation", err);
+            return reply.status(200).send("");
+          }
+        },
+      );
+
+      this.app.get<{ Params: { id: string }; Reply: string }>(
+        bot.getPath(":id"),
+        async (req, reply) => {
+          const routeId = req.params.id;
+          const botId = bot.getId();
+          const isLatestRouteId = routeId === botId;
+          if (!isLatestRouteId) {
+            logger.warn(
+              "Wrong route id! Perhaps because of the cache on Telegram side.",
+              {
+                routeId,
+                botId,
+                method: req.method,
+                url: req.originalUrl,
+              },
+            );
+          }
+
+          return reply
+            .status(200)
+            .type("text/plain")
+            .send(
+              isLatestRouteId
+                ? "Route is enabled"
+                : "Route is enabled under new routeId",
+            );
+        },
+      );
+    });
+
+    return this;
+  }
+
+  public start(): Promise<VoidPromise> {
+    logger.info(
+      `Starting ${Logger.y(sSuffix("http", this.isHttps))} ${this.selfUrl} server`,
+    );
+
+    return new Promise((resolve) => {
+      this.app.listen({ port: this.port, host: "0.0.0.0" }, () => {
+        logger.info(`The bot server is listening on ${Logger.y(this.port)}`);
+        resolve(
+          () =>
+            new Promise((resolveFn, rejectFn) => {
+              logger.warn("Shutting down the server instance");
+              if (this.uptimeDaemon.isRunning) {
+                logger.warn("Stopping the daemon");
+                this.uptimeDaemon.stop();
+              }
+
+              this.app
+                .close()
+                .then(() => {
+                  logger.warn("The bot server has stopped");
+                  resolveFn();
+                })
+                .catch((err) => {
+                  logger.error("Unable to stop the bot server", err);
+                  rejectFn(err);
+                });
+            }),
+        );
+      });
     });
   }
 
@@ -267,9 +283,56 @@ export class BotServer
     });
   }
 
-  protected getServerInstance(): Express {
-    const app = express();
-    app.use(express.json());
-    return app;
+  private async getStatusHandler(): Promise<[number, HealthDto]> {
+    const status = new HealthModel(
+      this.version,
+      this.isHttps,
+      this.threadId,
+      this.serverName,
+      this.nodeVersion,
+    );
+    if (this.isIdle) {
+      status.setMessage("App is not connected to the Telegram server");
+      return [400, status.getDto()];
+    }
+
+    if (!this.stat?.isReady()) {
+      logger.warn("Database is not ready");
+      // TODO fix behavior
+      // const errMessage = "Database is not ready";
+      // logger.error(errMessage, new Error(errMessage));
+      // status.setMessage("Database is not ready");
+      // res.status(400).send(status.getDto());
+      // return;
+    }
+
+    const days = this.uptimeDaemon.getCurrent();
+    status.setDaysOnline(days.current, days.limit);
+
+    try {
+      const urls = await Promise.all(
+        this.bots.map((bot) => bot.getHostLocation()),
+      );
+      status.setOnline(urls);
+      return [200, status.getDto()];
+    } catch (err) {
+      const errMessage = "Unable to get bot urls";
+      logger.error(errMessage, err);
+      status.setMessage(errMessage, true);
+      return [400, status.getDto()];
+    }
+  }
+
+  protected getServerInstance(): FastifyInstance {
+    const httpsOpts = this.isHttps
+      ? {
+          https: this.httpsOptions,
+        }
+      : {};
+
+    return fastify({
+      ...httpsOpts,
+      bodyLimit: getMB(100), // 100 MB
+    });
   }
 }
